viewer/requirements.txt
yapf
flask
google-cloud-datastore
nose
rednose
